+++ internal/proxy/azure_user.go
@@ -1,4 +1,4 @@
-package azure
+package proxy
 
 import (
 	"context"
@@ -6,23 +6,22 @@ import (
 	"github.com/go-logr/logr"
 	hamiltonMsgraph "github.com/manicminer/hamilton/msgraph"
 	hamiltonOdata "github.com/manicminer/hamilton/odata"
-	"github.com/xenitab/azad-kube-proxy/internal/cache"
 	"github.com/xenitab/azad-kube-proxy/internal/models"
 )
 
-type user struct {
-	cacheClient cache.ClientInterface
+type azureUser struct {
+	cache       Cache
 	usersClient *hamiltonMsgraph.UsersClient
 }
 
-func newUser(ctx context.Context, cacheClient cache.ClientInterface, usersClient *hamiltonMsgraph.UsersClient) *user {
-	return &user{
-		cacheClient: cacheClient,
+func newUser(ctx context.Context, cacheClient Cache, usersClient *hamiltonMsgraph.UsersClient) *azureUser {
+	return &azureUser{
+		cache:       cacheClient,
 		usersClient: usersClient,
 	}
 }
 
-func (user *user) getGroups(ctx context.Context, objectID string) ([]models.Group, error) {
+func (user *azureUser) getGroups(ctx context.Context, objectID string) ([]models.Group, error) {
 	log := logr.FromContextOrDiscard(ctx)
 
 	odataQuery := hamiltonOdata.Query{}
@@ -35,7 +34,7 @@ func (user *user) getGroups(ctx context.Context, objectID string) ([]models.Grou
 
 	var groups []models.Group
 	for _, group := range *groupsResponse {
-		group, found, err := user.cacheClient.GetGroup(ctx, *group.ID())
+		group, found, err := user.cache.GetGroup(ctx, *group.ID())
 		if err != nil {
 			return nil, err
 		}
