+++ internal/proxy/cache.go
@@ -1,4 +1,4 @@
-package cache
+package proxy
 
 import (
 	"context"
@@ -6,8 +6,7 @@ import (
 	"github.com/xenitab/azad-kube-proxy/internal/models"
 )
 
-// ClientInterface ...
-type ClientInterface interface {
+type Cache interface {
 	GetUser(ctx context.Context, s string) (models.User, bool, error)
 	SetUser(ctx context.Context, s string, u models.User) error
 	GetGroup(ctx context.Context, s string) (models.Group, bool, error)
